+++ pyriemann/utils/covariance.py
@@ -1,5 +1,6 @@
 import numpy
 from sklearn.covariance import oas, ledoit_wolf, fast_mcd, empirical_covariance
+import warnings
 
 # Mapping different estimator on the sklearn toolbox
 
@@ -113,18 +114,39 @@ def eegtocov(sig, window=128, overlapp=0.5, padding=True, estimator='cov'):
     return numpy.array(X)
 
 
-def coherence(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None):
-    """Compute coherence."""
-    cosp = cospectrum(X, window, overlap, fmin, fmax, fs)
-    coh = numpy.zeros_like(cosp)
-    for f in range(cosp.shape[-1]):
-        psd = numpy.sqrt(numpy.diag(cosp[..., f]))
-        coh[..., f] = cosp[..., f] / numpy.outer(psd, psd)
-    return coh
+###############################################################################
 
 
-def cospectrum(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None):
-    """Compute Cospectrum."""
+def cross_spectrum(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None):
+    """Compute the complex cross-spectral matrices of a real signal X.
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_trials, n_channels, n_samples)
+        ndarray of trials.
+    window : int (default 128)
+        The length of the FFT window used for spectral estimation.
+    overlap : float (default 0.75)
+        The percentage of overlap between window.
+    fmin : float | None, (default None)
+        The minimal frequency to be returned.
+    fmax : float | None, (default None)
+        The maximal frequency to be returned.
+    fs : float | None, (default None)
+        The sampling frequency of the signal.
+
+    Returns
+    -------
+    S : ndarray, shape (n_trials, n_channels, n_channels, n_freqs)
+        ndarray of cross-spectral matrices for each trials and for each
+        frequency bin.
+    freqs : ndarray, shape (n_freqs,)
+        The frequencies associated to cospectra.
+
+    References
+    ----------
+    .. [1] https://en.wikipedia.org/wiki/Cross-spectrum
+    """
     Ne, Ns = X.shape
     number_freqs = int(window / 2)
 
@@ -151,21 +173,107 @@ def cospectrum(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None):
         fdata[window_ix, :, :] = numpy.fft.fft(
             cdata, n=window, axis=1)[:, 0:number_freqs]
 
-    # Adjust Frequency range to specified range (in case it is a parameter)
-    if fmin is not None:
+    # adjust frequency range to specified range (in case it is a parameter)
+    if fs is not None:
+        if fmin is None:
+            fmin = 0
+        if fmax is None:
+            fmax = fs / 2
+        if fmax <= fmin:
+            raise ValueError('Parameter fmax must be superior to fmin')
+        if 2.0 * fmax > fs: # check Nyquist-Shannon
+            raise ValueError('Parameter fmax must be inferior to fs/2')
         f = numpy.arange(0, 1, 1.0 / number_freqs) * (fs / 2.0)
         Fix = (f >= fmin) & (f <= fmax)
         fdata = fdata[:, :, Fix]
+        freqs = f[Fix]
+    else:
+        if fmin is not None:
+            warnings.warn('Parameter fmin not used because fs is None')
+        if fmax is not None:
+            warnings.warn('Parameter fmax not used because fs is None')
+        freqs = None
 
-    # fdata = fdata.real
     Nf = fdata.shape[2]
     S = numpy.zeros((Ne, Ne, Nf), dtype=complex)
-    normval = numpy.linalg.norm(win)**2
     for i in range(Nf):
-        S[:, :, i] = numpy.dot(fdata[:, :, i].conj().T, fdata[:, :, i]) / (
-            number_windows * normval)
+        S[:, :, i] = numpy.dot(fdata[:, :, i].conj().T, fdata[:, :, i])
+    S /= number_windows * numpy.linalg.norm(win)**2
 
-    return numpy.abs(S)**2
+    return S, freqs
+
+
+def cospectrum(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None):
+    """Compute co-spectral matrices, the real part of cross-spectra.
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_trials, n_channels, n_samples)
+        ndarray of trials.
+    window : int (default 128)
+        The length of the FFT window used for spectral estimation.
+    overlap : float (default 0.75)
+        The percentage of overlap between window.
+    fmin : float | None, (default None)
+        The minimal frequency to be returned.
+    fmax : float | None, (default None)
+        The maximal frequency to be returned.
+    fs : float | None, (default None)
+        The sampling frequency of the signal.
+
+    Returns
+    -------
+    S : ndarray, shape (n_trials, n_channels, n_channels, n_freqs)
+        ndarray of co-spectral matrices for each trials and for each
+        frequency bin.
+    freqs : ndarray, shape (n_freqs,)
+        The frequencies associated to cospectra.
+    """
+    S, freqs = cross_spectrum(
+        X=X,
+        window=window,
+        overlap=overlap,
+        fmin=fmin,
+        fmax=fmax,
+        fs=fs)
+
+    return S.real, freqs
+
+
+def coherence(X, window=128, overlap=0.75, fmin=None, fmax=None, fs=None):
+    """Compute coherence.
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_trials, n_channels, n_samples)
+        ndarray of trials.
+    window : int (default 128)
+        The length of the FFT window used for spectral estimation.
+    overlap : float (default 0.75)
+        The percentage of overlap between window.
+    fmin : float | None, (default None)
+        The minimal frequency to be returned.
+    fmax : float | None, (default None)
+        The maximal frequency to be returned.
+    fs : float | None, (default None)
+        The sampling frequency of the signal.
+
+    Returns
+    -------
+    C : ndarray, shape (n_trials, n_channels, n_channels, n_freqs)
+        ndarray of coherence matrices for each trials and for each
+        frequency bin.
+    """
+    S, _ = cross_spectrum(X, window, overlap, fmin, fmax, fs)
+    S2 = numpy.abs(S)**2 # squared cross-spectral modulus
+    C = numpy.zeros_like(S2)
+    for f in range(S2.shape[-1]):
+        psd = numpy.sqrt(numpy.diag(S2[..., f]))
+        C[..., f] = S2[..., f] / numpy.outer(psd, psd)
+    return C
+
+
+###############################################################################
 
 
 def normalize(X, norm):